Update copr repo and image annotation
This installs a crun that has both wasmedge and wasmtime enabled so we
also update the annotation to specify the specific handler we want i.e.
wasmtime, otherwise the old annotation would use whatever WASM runtime
was compiled in and crun defaults to wasmedge first in the list of WASM
runtimes that have been compiled in.